requirements.txt
aiohttp==3.10.5
altair==5.4.1
faiss-cpu==1.9.0
google-search-results==2.4.2
langchain==0.3.3
langchain-community==0.3.2
langchain-experimental==0.3.2
langchain-openai==0.2.2
numexpr==2.10.1
numpy==1.26.4
openai==1.53.0
pandas==2.2.2
pathlib==1.0.1
protobuf==5.28.2
pydantic==2.9.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
pyyaml==6.0.2
requests==2.32.3
sqlalchemy==2.0.35
streamlit==1.39.0
tenacity==8.5.0
tiktoken==0.8.0
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.1
openpyxl==3.1.5
st-weaviate-connection==0.1.0
weaviate==0.1.2
dspy==2.5.31

